Fix up MPLS and QinQ support.
Among other things, added some macros to dedup some code in the indexer.

+++ integration_test/test.sh
@@ -17,6 +17,7 @@
 DUMMY="${DUMMY-dummy0}"
 PORT="${PORT-9123}"
 BASEDIR="${BASEDIR-/tmp}"
+SKIP_CLEANUP="${SKIP_CLEANUP}"
 
 set -e
 cd $(dirname $0)
@@ -76,20 +77,22 @@ set +e
 STENOGRAPHER_PID=""
 STENOTYPE_PID=""
 function CleanUp {
-  Info "Cleaning up"
-  if [ ! -z "$STENOGRAPHER_PID" ]; then
-    Info "Killing stenographer ($STENOGRAPHER_PID)"
-    KILLCMD=kill ReallyKill $STENOGRAPHER_PID
+  if [ -z "$SKIP_CLEANUP" ]; then
+    Info "Cleaning up"
+    if [ ! -z "$STENOGRAPHER_PID" ]; then
+      Info "Killing stenographer ($STENOGRAPHER_PID)"
+      KILLCMD=kill ReallyKill $STENOGRAPHER_PID
+    fi
+    if [ ! -z "$STENOTYPE_PID" ]; then
+      Info "Killing stenotype ($STENOTYPE_PID)"
+      KILLCMD=kill ReallyKill $STENOTYPE_PID
+    fi
+    Info "Deleting $DUMMY interface"
+    Info "Removing $OUTDIR"
+    rm -rfv $OUTDIR
+    sudo ifconfig $DUMMY down
+    sudo ip link del dummy0
   fi
-  if [ ! -z "$STENOTYPE_PID" ]; then
-    Info "Killing stenotype ($STENOTYPE_PID)"
-    KILLCMD=kill ReallyKill $STENOTYPE_PID
-  fi
-  Info "Deleting $DUMMY interface"
-  Info "Removing $OUTDIR"
-  rm -rfv $OUTDIR
-  sudo ifconfig $DUMMY down
-  sudo ip link del dummy0
 }
 trap CleanUp EXIT
